Change ets lookup to case statement.
This is because there is a case where a table can be created, but the lookup would still have returned "false" if the user was using OTP < 20. This can lead to a crash because if the user tries to create the table again (because they were told it did not exist), they will get an ArgumentError, which will then cause a crash.